requirements.txt
fastapi
uvicorn
google-generativeai
python-docx